pyproject.toml
[tool.poetry]
name = "cognito-pyauth"
version = "1.0.1"
description = ""
readme = "README.md"
authors = ["mountain-gorilla-inc"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
requests = "^2.26.0"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
pydantic = "^1.9.0"
fastapi = "^0.70.1"
boto3 = "^1.21.6"

[tool.poetry.dev-dependencies]
mypy = "^0.930"
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
invoke = "^1.6.0"
types-requests = "^2.26.3"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
python-dotenv = "^0.19.2"

[tool.mypy]
python_version = 3.9
plugins = "pydantic.mypy"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
